# app/static/swagger.json
{
    "openapi": "3.0.0",
    "info": {
      "title": "System Rejestracji Pacjenta Stomatologicznego API",
      "description": "API dla systemu rejestracji pacjenta stomatologicznego",
      "version": "1.0.0"
    },
    "servers": [
      {
        "url": "/api",
        "description": "Development server"
      }
    ],
    "components": {
      "securitySchemes": {
        "jwt": {
          "type": "http",
          "scheme": "bearer",
          "bearerFormat": "JWT"
        }
      }
    },
    "paths": {
...
      "/dokumentacja/diagram/{pacjent_id}": {
        "get": {
          "tags": ["Dokumentacja"],
          "summary": "Pobieranie diagramu dentystycznego pacjenta",
          "security": [
            { "jwt": [] }
          ],
          "parameters": [
            {
              "name": "pacjent_id",
              "in": "path",
              "required": true,
              "schema": {
                "type": "integer"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Diagram dentystyczny pacjenta"
            },
            "404": {
              "description": "Pacjent nie istnieje"
            }
          }
        }
      },
      "/dokumentacja/recepta/{wizyta_id}": {
        "post": {
          "tags": ["Dokumentacja"],
          "summary": "Wystawienie e-Recepty",
          "security": [
            { "jwt": [] }
          ],
          "parameters": [
            {
              "name": "wizyta_id",
              "in": "path",
              "required": true,
              "schema": {
                "type": "integer"
              }
            }
          ],
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "opis": {"type": "string"},
                    "leki": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "nazwa": {"type": "string"},
                          "dawkowanie": {"type": "string"},
                          "ilosc": {"type": "integer"},
                          "refundacja": {"type": "boolean"}
                        }
                      }
                    }
                  },
                  "required": ["opis", "leki"]
                }
              }
            }
          },
          "responses": {
            "201": {
              "description": "E-Recepta została wystawiona pomyślnie"
            },
            "404": {
              "description": "Wizyta nie istnieje"
            }
          }
        }
      },
...
    }
  }